chore: moved api calls to transport module
