+++ cmd/raid.go
@@ -1,7 +1,8 @@
 package cmd
 
 import (
-	"github.com/privateerproj/privateer-raid-osps-baseline/armory"
+	"github.com/eddie-knight/raid-osps-baseline/armory"
+	"github.com/eddie-knight/raid-osps-baseline/transport"
 
 	"github.com/privateerproj/privateer-sdk/config"
 	"github.com/privateerproj/privateer-sdk/raidengine"
@@ -29,7 +30,7 @@ func (r *Raid) Start() (err error) {
 
 func initializer(c *config.Config) (err error) {
 	armory.Config = c // for strikes to reference. TODO: not sure yet whether this mitigates the need for armory.Armory.Config
-	repoData, err := getRepositoryData(c)
+	repoData, err := transport.GetRepoData(c)
 	if err != nil {
 		return
 	}
